dX_in to inceptorData
PWM inputs from the RC receiver were in controlData.dX_in where X was
the control surface. These were moved to sensorData.inceptorData with
the signal names throttle, pitch, roll, and yaw. The data acquisition
was modified to reflect this change. Additionally, all of the aircraft
datalog's were modified to look for and log the new signal names. In the
aircraft configuration files, signal names for the channels and
calibration were changed to match the new inceptor_axis format rather
than pwmin_surface format that was used previously. These modifications
were made to increase readability in the signal naming conventions and
store the signals in the more appropriate data structure.