Correct paths and control_SIL.c function calls
The path to various files no longer needs the 'Software' folder.  Paths
in SIL have been updated to reflect this and resolve broken paths.

The c-code restructure (introduction of mission and inceptor structs)
changes the function arguments for control and guidance.  The SIL has
been updated to supply proper arguments.